requirements.txt
flask==3.0.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
scrapy==2.11.2
itemadapter==0.9.0
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
